+++ contracts/DepositManager.sol
@@ -1,7 +1,7 @@
 pragma solidity ^0.5.15;
 pragma experimental ABIEncoderV2;
 import { Types } from "./libs/Types.sol";
-import { Logger } from "./logger.sol";
+import { Logger } from "./Logger.sol";
 import { RollupUtils } from "./libs/RollupUtils.sol";
 import { MerkleTreeUtils as MTUtils } from "./MerkleTreeUtils.sol";
 import { NameRegistry as Registry } from "./NameRegistry.sol";
@@ -10,7 +10,7 @@ import { IERC20 } from "./interfaces/IERC20.sol";
 import { ParamManager } from "./libs/ParamManager.sol";
 import { POB } from "./POB.sol";
 import { Governance } from "./Governance.sol";
-import { Rollup } from "./rollup.sol";
+import { Rollup } from "./Rollup.sol";
 
 contract DepositManager {
     MTUtils public merkleUtils;
