+++ contracts/TokenRegistry.sol
@@ -1,6 +1,6 @@
 pragma solidity >=0.4.21;
 
-import { Logger } from "./logger.sol";
+import { Logger } from "./Logger.sol";
 import { NameRegistry as Registry } from "./NameRegistry.sol";
 import { ParamManager } from "./libs/ParamManager.sol";
 import { POB } from "./POB.sol";
