+++ contracts/prod/Airdrop.sol
@@ -1,6 +1,6 @@
 pragma solidity ^0.5.15;
 pragma experimental ABIEncoderV2;
 
-import { Airdrop } from "../airdrop.sol";
+import { Airdrop } from "../Airdrop.sol";
 
 contract AirdropProduction is Airdrop {}
