+++ test/merkleTree.test.ts
